feat(*): add rule management
